fix some origin bugs
