+++ be/src/vec/core/column_with_type_and_name.h
@@ -65,7 +65,7 @@ struct ColumnWithTypeAndName {
 
     void to_pb_column_meta(PColumnMeta* col_meta) const;
 
-    ColumnWithTypeAndName get_nested(bool replace_null_data_to_default = false) const;
+    ColumnWithTypeAndName get_nested_of_nullable(bool replace_null_data_to_default = false) const;
 
     Status check_type_and_column_match() const;
 };
