+++ be/src/vec/functions/date_time_transforms.h
@@ -403,91 +403,6 @@ struct FromUnixTimeDecimalImpl {
     }
 };
 
-template <PrimitiveType FromPType, PrimitiveType ToPType, typename Transform>
-struct Transformer {
-    using FromType = typename PrimitiveTypeTraits<FromPType>::ColumnItemType;
-    using ToType = typename PrimitiveTypeTraits<ToPType>::ColumnItemType;
-    static void vector(const PaddedPODArray<FromType>& vec_from, PaddedPODArray<ToType>& vec_to,
-                       NullMap& null_map) {
-        size_t size = vec_from.size();
-        vec_to.resize(size);
-        null_map.resize(size);
-
-        for (size_t i = 0; i < size; ++i) {
-            // The transform result maybe an int, int32, but the result maybe short
-            // for example, year function. It is only a short.
-            auto res = Transform::execute(vec_from[i]);
-            using RESULT_TYPE = std::decay_t<decltype(res)>;
-            vec_to[i] = cast_set<ToType, RESULT_TYPE, false>(res);
-        }
-
-        for (size_t i = 0; i < size; ++i) {
-            if (!((typename PrimitiveTypeTraits<Transform::OpArgType>::CppType&)(vec_from[i]))
-                         .is_valid_date()) [[unlikely]] {
-                throw_out_of_bound_one<typename PrimitiveTypeTraits<Transform::OpArgType>::CppType>(
-                        Transform::name, vec_from[i]);
-            }
-        }
-    }
-
-    static void vector(const PaddedPODArray<FromType>& vec_from, PaddedPODArray<ToType>& vec_to) {
-        size_t size = vec_from.size();
-        vec_to.resize(size);
-
-        for (size_t i = 0; i < size; ++i) {
-            auto res = Transform::execute(vec_from[i]);
-            using RESULT_TYPE = std::decay_t<decltype(res)>;
-            vec_to[i] = cast_set<ToType, RESULT_TYPE, false>(res);
-            DCHECK(((typename PrimitiveTypeTraits<Transform::OpArgType>::CppType&)(vec_from[i]))
-                           .is_valid_date());
-        }
-    }
-};
-
-template <PrimitiveType FromPType, PrimitiveType ToPType, template <PrimitiveType> typename Impl>
-struct TransformerYear {
-    using FromType = typename PrimitiveTypeTraits<FromPType>::ColumnItemType;
-    using ToType = typename PrimitiveTypeTraits<ToPType>::ColumnItemType;
-    static void vector(const PaddedPODArray<FromType>& vec_from, PaddedPODArray<ToType>& vec_to,
-                       NullMap& null_map) {
-        size_t size = vec_from.size();
-        vec_to.resize(size);
-        null_map.resize(size);
-
-        auto* __restrict to_ptr = vec_to.data();
-        auto* __restrict from_ptr = vec_from.data();
-        auto* __restrict null_map_ptr = null_map.data();
-
-        for (size_t i = 0; i < size; ++i) {
-            to_ptr[i] = Impl<FromPType>::execute(from_ptr[i]);
-        }
-
-        for (size_t i = 0; i < size; ++i) {
-            null_map_ptr[i] = to_ptr[i] > MAX_YEAR;
-        }
-    }
-
-    static void vector(const PaddedPODArray<FromType>& vec_from, PaddedPODArray<ToType>& vec_to) {
-        size_t size = vec_from.size();
-        vec_to.resize(size);
-
-        auto* __restrict to_ptr = vec_to.data();
-        auto* __restrict from_ptr = vec_from.data();
-
-        for (size_t i = 0; i < size; ++i) {
-            to_ptr[i] = Impl<FromPType>::execute(from_ptr[i]);
-        }
-    }
-};
-
-template <PrimitiveType FromType, PrimitiveType ToType>
-struct Transformer<FromType, ToType, ToYearImpl<FromType>>
-        : public TransformerYear<FromType, ToType, ToYearImpl> {};
-
-template <PrimitiveType FromType, PrimitiveType ToType>
-struct Transformer<FromType, ToType, ToYearOfWeekImpl<FromType>>
-        : public TransformerYear<FromType, ToType, ToYearOfWeekImpl> {};
-
 #include "common/compile_check_end.h"
 } // namespace doris::vectorized
