+++ be/test/vec/core/column_with_type_and_name_test.cpp
@@ -20,9 +20,10 @@
 #include <gtest/gtest.h>
 
 #include "testutil/column_helper.h"
+#include "vec/columns/column.h"
+#include "vec/columns/column_const.h"
 #include "vec/core/types.h"
 #include "vec/data_types/data_type_nullable.h"
-#include "vec/data_types/data_type_number.h"
 
 namespace doris::vectorized {
 
@@ -34,7 +35,8 @@ TEST(ColumnWithTypeAndNameTest, get_nested_test) {
     column_with_type_and_name.type =
             std::make_shared<DataTypeNullable>(std::make_shared<DataTypeInt32>());
     column_with_type_and_name.name = "column_with_type_and_name";
-    column_with_type_and_name.get_nested(true);
+    auto result = column_with_type_and_name.get_nested_of_nullable(true);
+    EXPECT_TRUE(is_column_const(*result.column));
 }
 
 } // namespace doris::vectorized
